// src/components/game/board/board.type.ts
import * as THREE from 'three';
import { GLTF } from 'three-stdlib';

export type GLTFResult = GLTF & {
  nodes: {
    Platform: THREE.Mesh;
    Block000: THREE.Mesh;
    Block001: THREE.Mesh;
    Block002: THREE.Mesh;
    Block003: THREE.Mesh;
    Block004: THREE.Mesh;
    Block005: THREE.Mesh;
    Block006: THREE.Mesh;
    Block007: THREE.Mesh;
    Block008: THREE.Mesh;
    Block009: THREE.Mesh;
    Block010: THREE.Mesh;
    Block011: THREE.Mesh;
    Block012: THREE.Mesh;
    Block013: THREE.Mesh;
    Block014: THREE.Mesh;
    Block015: THREE.Mesh;
    Block016: THREE.Mesh;
    Block017: THREE.Mesh;
    Block018: THREE.Mesh;
    Block019: THREE.Mesh;
    Block020: THREE.Mesh;
    Block021: THREE.Mesh;
    Block022: THREE.Mesh;
    Block023: THREE.Mesh;
    Block024: THREE.Mesh;
    Block025: THREE.Mesh;
    Block026: THREE.Mesh;
    Block027: THREE.Mesh;
    Block028: THREE.Mesh;
    Block029: THREE.Mesh;
    Block030: THREE.Mesh;
    Block031: THREE.Mesh;
    Block032: THREE.Mesh;
    Block033: THREE.Mesh;
    Block034: THREE.Mesh;
    Block035: THREE.Mesh;
    Block036: THREE.Mesh;
    Block037: THREE.Mesh;
    Block038: THREE.Mesh;
    Block039: THREE.Mesh;
    Block040: THREE.Mesh;
    Block041: THREE.Mesh;
    Block042: THREE.Mesh;
    Block043: THREE.Mesh;
    Block044: THREE.Mesh;
    Block045: THREE.Mesh;
    Block046: THREE.Mesh;
    Block047: THREE.Mesh;
    Block048: THREE.Mesh;
    Block049: THREE.Mesh;
    Block050: THREE.Mesh;
    Block051: THREE.Mesh;
    Block052: THREE.Mesh;
    Block053: THREE.Mesh;
    Block054: THREE.Mesh;
    Block055: THREE.Mesh;
    Block056: THREE.Mesh;
    Block057: THREE.Mesh;
    Block058: THREE.Mesh;
    Block059: THREE.Mesh;
    Block060: THREE.Mesh;
    Block061: THREE.Mesh;
    Block062: THREE.Mesh;
    Block063: THREE.Mesh;
    Block064: THREE.Mesh;
    Block065: THREE.Mesh;
    Block066: THREE.Mesh;
    Block067: THREE.Mesh;
    Block068: THREE.Mesh;
    Block069: THREE.Mesh;
    Block070: THREE.Mesh;
    Block071: THREE.Mesh;
    Block072: THREE.Mesh;
    Block073: THREE.Mesh;
    Block074: THREE.Mesh;
    Block075: THREE.Mesh;
    Block076: THREE.Mesh;
    Block077: THREE.Mesh;
    Block078: THREE.Mesh;
    Block079: THREE.Mesh;
    Block080: THREE.Mesh;
    Wall000: THREE.Mesh;
    Wall001: THREE.Mesh;
    Wall002: THREE.Mesh;
    Wall003: THREE.Mesh;
    Wall004: THREE.Mesh;
    Wall005: THREE.Mesh;
    Wall006: THREE.Mesh;
    Wall007: THREE.Mesh;
    Wall008: THREE.Mesh;
    Wall009: THREE.Mesh;
    Wall010: THREE.Mesh;
    Wall011: THREE.Mesh;
    Wall012: THREE.Mesh;
    Wall013: THREE.Mesh;
    Wall014: THREE.Mesh;
    Wall015: THREE.Mesh;
    Wall016: THREE.Mesh;
    Wall017: THREE.Mesh;
    Wall018: THREE.Mesh;
    Wall019: THREE.Mesh;
    Plate001: THREE.Mesh;
    Plate000: THREE.Mesh;
    Pawn000: THREE.Mesh;
    Pawn001: THREE.Mesh;
    Container000: THREE.Mesh;
    Container001: THREE.Mesh;
  };
  materials: {
    PlatformMaterial: THREE.MeshStandardMaterial;
    BlockMaterial: THREE.MeshStandardMaterial;
    WallWhiteMaterial: THREE.MeshStandardMaterial;
    ContainerMaterial: THREE.MeshStandardMaterial;
    WallBlackMaterial: THREE.MeshStandardMaterial;
    PawnWhiteMaterial: THREE.MeshStandardMaterial;
    PawnBlackMaterial: THREE.MeshStandardMaterial;
  };
};
export type Nodes = GLTFResult['nodes'];